Refactor create_rental
